fix(review): ella, doey 리뷰 반영  - backup_recovery_screen.dart 이름 수정  - 배경 CoconutColors.white 처리  - valut_list_restoration_view_model 추가  - 메소드명 수정(_buildWalletListItem)  - 복원 정보 화면과 볼트 리스트 복원 화면 분리
