+++ lib/app.dart
@@ -11,7 +11,8 @@ import 'package:coconut_vault/screens/airgap/psbt_confirmation_screen.dart';
 import 'package:coconut_vault/screens/airgap/psbt_scanner_screen.dart';
 import 'package:coconut_vault/screens/airgap/signed_transaction_qr_screen.dart';
 import 'package:coconut_vault/screens/airgap/single_sig_sign_screen.dart';
-import 'package:coconut_vault/screens/backup_recovery/backup_recovery_screen.dart';
+import 'package:coconut_vault/screens/backup_recovery/restoration_info_screen.dart';
+import 'package:coconut_vault/screens/backup_recovery/vault_list_restoration_screen.dart';
 import 'package:coconut_vault/screens/home/vault_list_screen.dart';
 import 'package:coconut_vault/screens/backup_recovery/prepare_update_screen.dart';
 import 'package:coconut_vault/screens/vault_creation/single_sig/security_self_check_screen.dart';
@@ -239,7 +240,10 @@ class _CoconutVaultAppState extends State<CoconutVaultApp> {
             AppRoutes.prepareUpdate: (context) => const CustomLoadingOverlay(
                   child: PrepareUpdateScreen(),
                 ),
-            AppRoutes.backupRecovery: (context) => const BackupRecoveryScreen(),
+            AppRoutes.restorationInfo: (context) =>
+                const RestorationInfoScreen(),
+            AppRoutes.vaultListRestoration: (context) =>
+                const VaultListRestorationScreen(),
           },
         ),
       ),
