+++ lib/constants/app_routes.dart
@@ -26,5 +26,6 @@ class AppRoutes {
   static const String welcome = '/welcome';
   static const String connectivityGuide = '/connectivity-guide';
   static const String prepareUpdate = '/prepare-update';
-  static const String backupRecovery = "/backup-recovery";
+  static const String restorationInfo = "/restoration-info";
+  static const String vaultListRestoration = "/vault-list-restoration";
 }
